Is this R object a RWR on Multiplex-Heterogeneous object (Results of the RWR-MH)?

A RWR on Multiplex Heterogeneous object is an R object generated as the result of calling the function Random.Walk.Restart.MultiplexHet (Results of the RWR-MH)

isRWRMH_Results(x)

Arguments

x
An R object

Value

A logical constant, TRUE if argument x is a RWRMH_Results object.

Details

isRWRMH_Results(x) checks whether an R object is RWRMH_Results

See also

Examples

m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject_1 <- create.multiplex(list(m1=m1,m2=m2))
h1 <- igraph::graph(c("A","C","B","E","E","D","E","C"), directed = FALSE)
multiObject_2 <- create.multiplex(list(h1=h1))
bipartite_relations <- data.frame(m=c(1,3),h=c("A","E"))
multiHetObject <-
    create.multiplexHet(multiObject_1,multiObject_2,bipartite_relations)

#> checking input arguments...

#> Generating bipartite matrix...

#> Expanding bipartite matrix to fit the multiplex network...

MultiHetTranMatrix <- compute.transition.matrix(multiHetObject)

#> Computing adjacency matrix of the first Multiplex network...

#> Computing adjacency matrix of the second Multiplex network...

#> Computing inter-subnetworks transitions...

#> Computing intra-subnetworks transitions...

#> Combining inter e intra layer probabilities into the global 
#>           Transition Matix

Multiplex1_Seeds <- c(1)
Multiplex2_Seeds <- c("E")
RWR_MultiHetResults <-
    Random.Walk.Restart.MultiplexHet(MultiHetTranMatrix,multiHetObject,
    Multiplex1_Seeds ,Multiplex2_Seeds)
isRWRMH_Results(RWR_MultiHetResults)

#> [1] TRUE

isRWRMH_Results(m1)

#> [1] FALSE