Create multiplex graphs from individual networks

create.multiplex is a function to create a multiplex network (Multiplex object) from up to 6 individual networks defined as igraph objects. See more details about multiplex networks below. If just one network is provided, a Multiplex object with one layer is therefore created (A monoplex network).

create.multiplex(...)

# S3 method for default
create.multiplex(L1, L2 = NULL, L3 = NULL, L4 = NULL,
  L5 = NULL, L6 = NULL, Layers_Name, ...)

Arguments

...
Further arguments passed to create.multiplex

L1
An igraph object describing a monoplex network. It will be integrated as the first layer of the multiplex network.

L2
An igraph object describing a monoplex network. It will be integrated as the second layer of the multiplex network. It's optional.

L3
An igraph object describing a monoplex network. It will be integrated as the third layer of the multiplex network. It's optional.

L4
An igraph object describing a monoplex network. It will be integrated as the fourth layer of the multiplex network. It's optional.

L5
An igraph object describing a monoplex network. It will be integrated as the fifth layer of the multiplex network. It's optional.

L6
An igraph object describing a monoplex network. It will be integrated as the sixth layer of the multiplex network. It's optional.

Layers_Name
A vector containing the names of the different layers. This name will be included as an attribute for all the edges of each network. It's optional. See more details below.

Value

A Multiplex object. It contains a list of the different graphs integrating the multiplex network, the names and number of its nodes and the number of layers.

Details

A multiplex network is a collection of layers (monoplex networks) sharing the same nodes, but in which the edges represent relationships of different nature. The number of layers of a multiplex object can vary from 1 (monoplex network) up to 6. Therefore, only the first layer is mandatory. We have limited the number of layers to 6 in order to reduce computation times for very large networks.

Layers_Name: A vector contianing the name of the different layers. It's optional, but if provided the number of layers should match the length of this vector. Its elements should be in the same order than the graphs (Layers_Name = c(L1_name, L2_name, L3_name, L4_name, L5_name, L6_name)

See also

Author

Alberto Valdeolivas Urbelz alvaldeolivas@gmail.com

Examples

m1 <- igraph::graph(c(1,2,1,3,2,3), directed = FALSE)
m2 <- igraph::graph(c(1,3,2,3,3,4,1,4), directed = FALSE)
multiObject <- create.multiplex(m1,m2)

#> Error in create.multiplex.default(m1, m2): The input object should be a list of graphs.